+++ readium/navigator/r2-navigator/src/main/java/org/readium/r2/navigator/pdf/PdfNavigatorFragment.kt
@@ -154,8 +154,10 @@ class PdfNavigatorFragment internal constructor(
     private val _currentLocator = MutableStateFlow(initialLocator ?: publication.readingOrder.first().toLocator())
 
     override fun go(locator: Locator, animated: Boolean, completion: () -> Unit): Boolean {
-        val page = ((locator.locations.page ?: 1) - 1).coerceAtLeast(0)
-        return goToHref(locator.href, page, animated, completion)
+        // FIXME: `position` is relative to the full publication, which would cause an issue for a publication containing several PDFs resources. Only publications with a single PDF resource are supported at the moment, so we're fine.
+        val pageNumber = locator.locations.page ?: locator.locations.position ?: 1
+        val pageIndex = (pageNumber - 1).coerceAtLeast(0)
+        return goToHref(locator.href, pageIndex, animated, completion)
     }
 
     override fun go(link: Link, animated: Boolean, completion: () -> Unit): Boolean =